requirements.txt
Django==5.0
requests
PyYAML
django-cors-headers
django-crispy-forms
django-filter
django-rest-auth
django-allauth
uritemplate
Markdown
Pygments
django-guardian
django-oauth-toolkit
djangorestframework
django-rest-swagger
django-rest-framework-social-oauth2
django-rest-framework-camel-case
django-rest-auth[with_social]
Pillow
psycopg2-binary
djangorestframework-simplejwt
django-dotenv
django-otp
pyotp
python-decouple